Test fn cargo_editor rather than the current environment
We would like these tests to pass no matter what the person who is
running the tests has set up their environment to be. We also don't want
to test that rust's std::env code is working; the behavior we'd like to
test here is that we're preferring the expected environment variable.

These tests are mostly currently failing since we haven't yet
incorporated a way to *set* the environment for the tests yet :)